10-queues/flask/requirements.txt
flask==1.1.2
beanstalkc
redis==3.5.3
python-dateutil==2.4.1